css/css-backgrounds/parsing/border-image-repeat-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-image-repeat with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-repeat">
<meta name="assert" content="border-image-repeat supports the full grammar '[ stretch | repeat | round | space ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
// "stretch" in Blink, Edge, WebKit; "stretch stretch" in Firefox.
test_valid_value("border-image-repeat", "stretch", ["stretch", "stretch stretch"]);
// "space" in Blink, Edge, WebKit; "space space" in Firefox.
test_valid_value("border-image-repeat", "space space", ["space", "space space"]);
test_valid_value("border-image-repeat", "repeat round");
</script>
</body>
</html>
